pyproject.toml
[project]
name = "grimoire"
version = "0.1.0"
description = "Genesis Engine - AI-native story generation system"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "aiosqlite>=0.20.0",
    "litellm>=1.0.0",
    "pydantic>=2.0.0",
    "loguru>=0.7.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "sqlite-vec>=0.1.0",
    "sentence-transformers>=2.2.0",
    "numpy>=1.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["backend/tests"]
markers = [
    "llm: marks tests as requiring real LLM API (deselect with '-k \"not llm\"')",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.hatch.build.targets.wheel]
packages = ["backend"]